eslint

+++ part4/bloglist/controllers/blogs.js
@@ -1,16 +1,16 @@
 const blogsRouter = require('express').Router()
-const { request, response } = require('../app')
+//const { request, response } = require('../app')
 const Blog = require('../models/blog')
 const User = require('../models/user')
 const jwt = require('jsonwebtoken')
 
 blogsRouter.get('/', async (request, response, next) => {
   try{
-    const blogs = await Blog.find({}).populate('user',{username:1, name:1 ,id:1})
+    const blogs = await Blog.find({}).populate('user', { username:1, name:1 ,id:1 })
     response.json(blogs)
   }catch(exception){
     next(exception)
-  } 
+  }
 })
 
 blogsRouter.get('/:id', async (request, response, next) => {
@@ -21,10 +21,10 @@ blogsRouter.get('/:id', async (request, response, next) => {
     }else{
       response.status(404).end()
     }
-    
+
   }catch(exception){
     next(exception)
-  } 
+  }
 })
 
 // const getTokenFrom = request => {
@@ -39,13 +39,14 @@ blogsRouter.post('/', async (request, response, next) => {
   try{
     const body = request.body
 
-    //const token = getTokenFrom(request)
     const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    //console.log(decodedToken)
+
     if (!request.token || !decodedToken.id) {
       return response.status(401).json({ error: 'token missing or invalid' })
     }
     const user = await User.findById(decodedToken.id)
-  
+
     const newBlog = new Blog(
       {
         title: body.title ,
@@ -67,7 +68,22 @@ blogsRouter.post('/', async (request, response, next) => {
 })
 
 blogsRouter.delete('/:id', async (request, response, next) => {
+
   try{
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!request.token || !decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+    const user = await User.findById(decodedToken.id)
+
+    const blogsid = user.blogs.map(blogObjectId => blogObjectId.toString())
+    //console.log(blogsid)
+
+    if(!blogsid.includes(request.params.id)){
+      response.status(401).json({ Error: 'Unauthorized' }).end()
+    }
+
+
     const blog = await Blog.findByIdAndRemove(request.params.id)
     if(blog){
       response.status(204).end()
@@ -89,7 +105,7 @@ blogsRouter.put('/:id', async (request,response,next) => {
     }
 
 
-    const result = await Blog.findByIdAndUpdate(request.params.id, body , {new: true} )
+    const result = await Blog.findByIdAndUpdate(request.params.id, body , { new: true } )
 
     if(result){
       response.status(200).json(result)
